Revert "User Page Mock"
