--- .prettierrc
@@ -1,1 +1,0 @@
-{ "singleQuote": true }
