+++ package.json
@@ -10,9 +10,7 @@
     "node-sass": "^6.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^6.0.2",
     "react-scripts": "4.0.3",
-    "recharts": "^2.1.6",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
