--- src/App.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import { Route, Routes } from 'react-router-dom';
-import HomePage from './pages/HomePage';
-import UserPage from './pages/UserPage';
-
-function App() {
-  return (
-    <Routes>
-      <Route exact path="/" element={<HomePage />} />
-      <Route path="/user/:name" element={<UserPage />} />
-    </Routes>
-  );
-}
-
-export default App;
