+++ src/index.scss
@@ -31,7 +31,7 @@ img.profile {
   transform: translateY(6px);
 
   .icon {
-    background: rgba(255, 255, 255, 0.65);
+    background: rgba(255,255,255,0.65);
     position: absolute;
     left: 0;
     top: 0;
@@ -49,8 +49,7 @@ img.profile {
   }
 }
 
-td,
-th {
+td,th {
   white-space: nowrap;
   padding: 8px 16px !important;
 }
